moves smt solver statistics delta calculations to SymbExLogger
